Extract inspectVNode
